version 0.3.0.2

+++ FlounderLib/BitBoard.fs
@@ -67,7 +67,7 @@ type BitBoard =
         member this.ToUint64() =
             this.Internal
         member this.ToSq():Square =
-            LanguagePrimitives.EnumOfValue(sbyte(BitOperations.TrailingZeroCount(this.Internal)))
+            LanguagePrimitives.EnumOfValue(BitOperations.TrailingZeroCount(this.Internal))
         member this.ToSqs():Square array =
             let c = this.Count
             let mutable iterator = BitBoardIterator(this.Internal, c)
